Rename pkg to cortex
